Fix deadlock issues with capturing stdout from process

+++ IronScheme/IronScheme.Tests/TestRunner.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using NUnit.Framework;
 using System.Diagnostics;
+using System.IO;
 
 namespace IronScheme.Tests
 {
@@ -15,7 +16,17 @@ namespace IronScheme.Tests
       public int ExitCode;
     }
 
+    protected TestResult RunTestWithInput(string input)
+    {
+      return RunTest(null, input);
+    }
+
     protected TestResult RunTest(string filename)
+    {
+      return RunTest(filename, null);
+    }
+
+    protected TestResult RunTest(string filename, string input)
     {
       var p = new Process
       {
@@ -27,28 +38,50 @@ namespace IronScheme.Tests
           RedirectStandardInput = true,
           UseShellExecute = false,
           CreateNoWindow = true,
-          Arguments = filename
+          Arguments = filename,
+        },
+        //EnableRaisingEvents = true,
+      };
+
+      //p.OutputDataReceived += (s, e) => Console.Write(e.Data);
+
+      try
+      {
+        p.Start();
+
+        if (input != null)
+        {
+          p.StandardInput.WriteLine(input);
+          p.StandardInput.WriteLine("(exit)");
+          p.StandardInput.Flush();
+          p.StandardInput.Close();
         }
-      };
 
-      p.Start();
+        var r = new TestResult
+        {
+          Output = p.StandardOutput.ReadToEnd()
+        };
 
-      var exited = p.WaitForExit(90000);
+        var exited = p.WaitForExit(12000000);
 
-      if (!exited)
-      {
-        p.Kill();
+        if (!exited)
+        {
+          p.Kill();
+        } 
+
+        Assert.AreEqual(0, p.ExitCode);
+
+        return r;
       }
-
-      var r = new TestResult
+      catch (Exception ex)
       {
-        ExitCode = p.ExitCode,
-        Output = p.StandardOutput.ReadToEnd()
-      };
-
-      Assert.AreEqual(0, p.ExitCode);
-
-      return r;
+        if (!p.HasExited)
+        {
+          p.Kill();
+        }
+        Assert.Fail("runner failed: {0}", ex.Message);
+        throw;
+      }
     }
   }
 }
